requirements.txt
Pillow>=10.0.0
piexif>=1.1.3
pandas>=2.0.0
numpy>=1.24.0